M3: calculator split + recompute_windows trim
- Extract parse_schedule() and parse_period_rates() from calculator.rs
  into src/trueup/openei_parse.rs; PeriodRate stays in calculator.rs
  and is imported by openei_parse (pub(crate) struct)
- calculator.rs reduces from 277 lines to 220 (182 code_lines per scorer)
- Move run_typed() and run_raw() from recompute_windows binary into
  src/trueup/recompute.rs library module; shared apply_recompute() helper
  eliminates the ~40-line duplicated inner block
- recompute_windows.rs reduces from 257 lines to 41 (thin entry point:
  Args + main + dispatch)
- All 68 tests pass; log event set unchanged; clippy clean
