+++ src/trueup/calculator.rs
@@ -4,6 +4,8 @@ use chrono::{Datelike, TimeZone, Timelike};
 use chrono_tz::America::Los_Angeles;
 use std::collections::{BTreeSet, HashMap};
 
+use super::openei_parse::{parse_period_rates, parse_schedule};
+
 #[derive(Debug, Default, Clone)]
 pub struct PeriodSummary {
     pub import_kwh: f64,
@@ -27,9 +29,9 @@ enum TouPeriod {
     SuperOffPeak,
 }
 
-struct PeriodRate {
-    rate: f64,
-    sell_rate: f64,
+pub(crate) struct PeriodRate {
+    pub(crate) rate: f64,
+    pub(crate) sell_rate: f64,
 }
 
 pub fn calculate(
@@ -131,64 +133,6 @@ pub fn calculate(
         super_off_peak,
         net_cost_usd,
     })
-}
-
-fn parse_schedule(val: &serde_json::Value) -> Result<Vec<Vec<usize>>, AppError> {
-    let months = val.as_array().ok_or_else(|| {
-        AppError::Tou(TouError::ParseError(
-            "energyweekdayschedule is not an array".into(),
-        ))
-    })?;
-    months
-        .iter()
-        .map(|month_val| {
-            let hours = month_val.as_array().ok_or_else(|| {
-                AppError::Tou(TouError::ParseError(
-                    "schedule month is not an array".into(),
-                ))
-            })?;
-            hours
-                .iter()
-                .map(|h| {
-                    h.as_i64().map(|n| n as usize).ok_or_else(|| {
-                        AppError::Tou(TouError::ParseError("invalid period index".into()))
-                    })
-                })
-                .collect()
-        })
-        .collect()
-}
-
-fn parse_period_rates(val: &serde_json::Value) -> Result<Vec<PeriodRate>, AppError> {
-    let periods = val.as_array().ok_or_else(|| {
-        AppError::Tou(TouError::ParseError(
-            "energyratestructure is not an array".into(),
-        ))
-    })?;
-    periods
-        .iter()
-        .enumerate()
-        .map(|(i, period_val)| {
-            let tiers = period_val.as_array().ok_or_else(|| {
-                AppError::Tou(TouError::ParseError("rate period is not an array".into()))
-            })?;
-            let tier = tiers.first().ok_or_else(|| {
-                AppError::Tou(TouError::ParseError("rate period has no tiers".into()))
-            })?;
-            let rate = tier["rate"].as_f64().ok_or_else(|| {
-                AppError::Tou(TouError::ParseError(format!(
-                    "missing 'rate' field in tier {i}"
-                )))
-            })?;
-            let sell_rate = if let Some(s) = tier["sell"].as_f64() {
-                s
-            } else {
-                tracing::warn!(event = "tou_sell_rate_missing", tier = i);
-                rate
-            };
-            Ok(PeriodRate { rate, sell_rate })
-        })
-        .collect()
 }
 
 fn build_per_month_maps(
