+++ src/trueup/mod.rs
@@ -1,1 +1,3 @@
 pub mod calculator;
+pub(crate) mod openei_parse;
+pub mod recompute;
